backend/chabot/requirements.txt
streamlit
python-dotenv
openai
pandas
numpy
fastapi
uvicorn
pymongo
python-jose[cryptography]
passlib[bcrypt]
cryptography
pydantic
pydantic-settings
email-validator
httpx
requests
pyotp
aiosmtplib
python-multipart
razorpay